2013-07-04  Basile Starynkevitch  <basile@starynkevitch.net>
	{{First MELT bootstrap with classy frames, but optimized flavour dont work!}}
	* melt-runtime.h (MELT_HAVE_CLASSY_FRAME): Can be set to 1.
	(Melt_CallProtoFrame): Improve dbgcall_file print.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/melt-branch@200675 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/melt-runtime.h
@@ -2917,7 +2917,7 @@ public:
   ~Melt_CallProtoFrame() {
 #if MELT_HAVE_DEBUG
     if (MELT_UNLIKELY( _dbgcall_file_ != NULL)) {
-      fprintf (_dbgcall_file_, "- %s:%d\n",  _meltcf_dbgfile, _meltcf_dbgline);
+      fprintf (_dbgcall_file_, "- %s:%d\n",  _meltcf_dbgfile, (int)_meltcf_dbgline);
       fflush (_dbgcall_file_);
     }
 #endif /*MELT_HAVE_DEBUG*/
